services/threat-scoring/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pyyaml==6.0.2
requests==2.32.3
pyjwt==2.9.0
